+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.20.0] - 2021-04-27
 ### Added
 - New variable `hasMoreThanOne` in the Installments Renderer for ICU messages, aiding in the creation of custom messages.